Tom Bending discovered a problem with gravity on sink particles using gravxyzpstore() and MPI.  The gravxyzpstore() optimisation is only used when using iptintree=1 when sink particles have smaller timesteps than all gas particles.  In this case, on timesteps when only sinks are being evolved, the gravity acting on sinks from the gas is only calculated when at least one gas particle is being evolved.  For other timesteps, a stored value of the gravity from the gas is used.
This was all fine, except when combined with MPI. Prior to this
update the forces on sink particles from gas on other MPI
processes would usual be incorrect (when only timestepping sink
particles) because the elements from the gravxyzpstore() array
that were used would not necessarily contain the forces for that
(foreign) sink particle.  In general, this is very unlikely to
matter because these forces would likely be small, and they
were only used for timesteps when non gas particles were being
evolved (e.g. if there was a very tight sink particle binary).
Nevertheless, it was a bug.

The fix in densityiterate_gradh_MPI.F is not to use the
gravxyzpstore() approximation for non-local sink particles.
Instead, the gravity from that MPI process acting on the foreign
sink is calculated as usual.  This is not an optimal fix, but
it is very simple (just modifying two IFs).

At the same time, incorrect code involving gravxyzpstore() was
corrected in insulate_TREE_P.F, but this only mattered for
calculations involving particle injection.

Modified: densityiterate_gradh_MPI.F insulate_TREE_P.F