Modify mtree and revtree so that the node radii can be calculated more accurately.
The PARAMETER ileveltreeacc (set in igrape) sets how far to walk the tree to improve the node radius accuracy. The larger ileveltreeacc is, the more accurate the tree but the more time mtree and revtree will take. An accurate tree speeds up neighbour finding. ileveltreeacc of 4 seems to be the best trade of in tests so far, with 3 being a more conservative choice. However, there are some circumstances in which a value > 1 could slow the code down (e.g. a very deep timestep distribution combined with high idim). This optimisation is therfore off by default.

This was also an opportunity to remove several sections of repeated code from the revtree routines. There were if statements that determined whether to use an openMP loop or perform the same loop in serial. This has been replaced using an openMP region if statement which has the same outcome without repeating the code.