Improvement to the tree node level values.
The tree nodes' levels are no longer defined by the iteration in mtree on which they were created. Instead they are determined by the maximum distance through the tree to a leaf. This reduces the total number of levels slightly improving the openMP performance in the revtree routines. It requires the tree nodes to be re-ordered at the end of mtree.