Increased the number of parents to be updated by leapfrog revtree when doing a partial tree revision.
Particles that had derivatives calculated in the previous timestep -- but not this timestep -- will have had smoothing lengths updated since the previous tree update. This is because smoothing lengths are updated in densityiterate not step. The parents of such particles are now added to the listparent array. This only needs to happen if the (smallest) current timestep is shorter than the previous one.